Move utility methods in to utilities object
