+++ app/utils/Utilities.scala
@@ -2,6 +2,8 @@ package utils
 
 import java.io.File
 
+import com.google.common.base.Charsets
+import com.google.common.io.BaseEncoding
 import play.api.libs.json._
 
 /**
@@ -9,7 +11,14 @@ import play.api.libs.json._
  */
 object Utilities {
 
+  private val DELIMITER: String = "~"
+
   private def currentDirectory = new File(".").getCanonicalPath
+
+  def encodeBase64(str: Seq[String], deliminator: String = ":"): String =
+    BaseEncoding.base64.encode(str.mkString(deliminator).getBytes(Charsets.UTF_8))
+
+  def decodeBase64(str: String): String = new String(BaseEncoding.base64().decode(str), "UTF-8")
 
   def errAsJson(status: Int, code: String, msg: String, cause: String = "Not traced"): JsObject = {
     Json.obj(
@@ -19,6 +28,17 @@ object Utilities {
       "message_en" -> msg
     )
   }
+
+  def createEntRowKey(period: String, id: String): String = String.join(DELIMITER, period, id)
+
+  def createUnitLinksRowKey(period: String, id: String, unitType: Option[String]): String = {
+    unitType match {
+      case Some(u) => String.join(DELIMITER, period, id, u)
+      case None => String.join(DELIMITER, period, id, "*")
+    }
+  }
+
+  def createTableNameWithNameSpace(nameSpace: String, tableName: String): String = s"$nameSpace:$tableName"
 
   // ret: AnyVal
   def getElement(value: Any) = {
